Break up environment
